Moved Storage Function
